requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
langchain-text-splitters
azure-ai-documentintelligence
psycopg[binary]
pgvector
python-multipart
tavily-python
yfinance
pandas
mplfinance
pandas_ta
